test: Fix pump tests on mac with python setting CPATH
As noted in the comment, in src/compile.c, distcc refuses to distribute
via pump if CPATH is set. For reasons not known yet to me, python is
setting CPATH:

    $ printenv CPATH
    [empty]
    $ python3 -c 'import os; print(os.getenv("CPATH"))'
    /usr/local/include

This means that tests were _never_ running pump on maca; it just wasn't
exposed until ScanIncludes_Case was added (which needs to pump to get
any response).

Signed-off-by: Alex Karle <akarle@mathworks.com>